requirements/base.txt
django==1.11.6
psycopg2==2.7.1
Pillow==4.3.0
sorl-thumbnail==12.3
django-crispy-forms==1.6.1
django-flat-theme==1.1.4
whitenoise==3.3.1
openpyxl==2.4.8
requests==2.18.4
XlsxWriter==1.0.0
python-dateutil==2.6.1
django-simple-history==1.9.0
sqlparse==0.2.4
djangorestframework==3.7.0
djangorestframework-jsonapi==2.2.0
django-filter==1.0.4
git+ssh://git@git.minsa.gob.pe/oidt/minsalogin-client.git@develop#egg=minsalogin_client
git+ssh://git@git.minsa.gob.pe/oidt/mpi-client.git@develop#egg=mpi_client
git+ssh://git@git.minsa.gob.pe/oidt/cita_client.git@develop#egg=cita_client
drfdocs==0.0.11
